refactor: reorganize header and footer components into sections for better structure
