+++ src/styles/globals.css
@@ -127,5 +127,3 @@
     }
   }
 }
-
-/* todo: fix global view transistion trigger on load: NOT FIXED */
